# openapi.yaml
openapi: 3.0.3
info:
  title: TaskAPI
  description: Task Management API
  version: 0.0.1
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []

paths:
  /actuator/health:
    get:
      tags: [actuator]
      summary: Health check
      security: []
      responses:
        "200":
          description: OK
          content:
            application/vnd.spring-boot.actuator.v3+json:
              schema:
                type: object

  /api/auth/register:
    post:
      tags: [auth]
      summary: Register a user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAuthResponse"

  /api/auth/login:
    post:
      tags: [auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAuthResponse"

  /api/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAuthResponse"

  /api/auth/logout:
    post:
      tags: [auth]
      summary: Logout
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseString"

  /api/v1/tasks/admin-all:
    get:
      tags: [tasks]
      summary: List all tasks (admin)
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"

  /api/v1/tasks:
    get:
      tags: [tasks]
      summary: List tasks for current user
      description: Use the optional status or title query parameters to filter tasks.
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: title
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
    post:
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"

  /api/v1/tasks/{id}:
    get:
      tags: [tasks]
      summary: Get task by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
    patch:
      tags: [tasks]
      summary: Patch task by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskPatchRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
    delete:
      tags: [tasks]
      summary: Delete task by id (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content

  /api/v1/users:
    get:
      tags: [users]
      summary: List users (admin)
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseUserResponse"
    post:
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUserResponse"

  /api/v1/users/{id}:
    get:
      tags: [users]
      summary: Get user by id (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUserResponse"
    delete:
      tags: [users]
      summary: Delete user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content

  /api/v1/users/me:
    get:
      tags: [users]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUserResponse"
    patch:
      tags: [users]
      summary: Update current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUserResponse"

  /api/v1/users/admin/{id}:
    patch:
      tags: [users]
      summary: Promote user to admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUserResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
    size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
    sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        example: createdAt,desc

  schemas:
    TaskStatus:
      type: string
      enum: [TODO, IN_PROGRESS, DONE]

    AuthRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
        password:
          type: string
          format: password

    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8

    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        username:
          type: string
        role:
          type: string

    TaskCreateRequest:
      type: object
      required: [title, description, taskStatus]
      properties:
        title:
          type: string
        description:
          type: string
        taskStatus:
          $ref: "#/components/schemas/TaskStatus"

    TaskPatchRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        taskStatus:
          $ref: "#/components/schemas/TaskStatus"

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        taskStatus:
          $ref: "#/components/schemas/TaskStatus"
        userId:
          type: integer
          format: int64
        username:
          type: string
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    UserRequest:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 5
          maxLength: 15
        password:
          type: string
          format: password

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        username:
          type: string
        role:
          type: string

    PageResponseTaskResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/TaskResponse"
        page:
          type: integer
        size:
          type: integer
        totalPages:
          type: integer
        totalElements:
          type: integer
          format: int64
        last:
          type: boolean

    PageResponseUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        page:
          type: integer
        size:
          type: integer
        totalPages:
          type: integer
        totalElements:
          type: integer
          format: int64
        last:
          type: boolean

    ApiResponseTaskResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/TaskResponse"
        timeStamp:
          type: string

    ApiResponseUserResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/UserResponse"
        timeStamp:
          type: string

    ApiResponseAuthResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/AuthResponse"
        timeStamp:
          type: string

    ApiResponsePageResponseTaskResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/PageResponseTaskResponse"
        timeStamp:
          type: string

    ApiResponsePageResponseUserResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/PageResponseUserResponse"
        timeStamp:
          type: string

    ApiResponseString:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          type: string
        timeStamp:
          type: string